# systems/if_line/docs/api/story-path-authoring.openapi.yaml
openapi: 3.1.0
info:
  title: IF Line StoryPath Authoring API
  version: 1.0.0
  description: >-
    Replacement authoring and publication contract. This contract is mounted
    directly below /api. Legacy chapter-index authoring routes are not part of
    this API and receive no compatibility adapter.
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
...
  /chapter-revisions/{chapter_revision_id}/script-generations:
    parameters:
      - {$ref: '#/components/parameters/ChapterRevisionId'}
    post:
      tags: [Scripts]
      operationId: generateChapterScript
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GenerationRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
  /chapter-revisions/{chapter_revision_id}/script-revisions:
    parameters:
      - {$ref: '#/components/parameters/ChapterRevisionId'}
    get:
      tags: [Scripts]
      operationId: listScriptRevisions
      responses:
        '200':
          description: Scripts generated for one exact chapter revision.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ScriptRevision'}
  /chapter-revisions/{chapter_revision_id}/script-head:
    parameters:
      - {$ref: '#/components/parameters/ChapterRevisionId'}
    get:
      tags: [Scripts]
      operationId: getScriptHead
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Scripts]
      operationId: updateScriptHead
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeadUpdate'}
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /chapter-script-revisions/{script_revision_id}/resource-slots:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
    get:
      tags: [Scripts]
      operationId: listScriptResourceSlots
      responses:
        '200':
          description: Resource slots for an immutable Script IR.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ResourceSlot'}
  /chapter-script-revisions/{script_revision_id}/resource-renders:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
    post:
      tags: [Scripts]
      operationId: renderScriptResources
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResourceRenderRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
  /chapter-script-revisions/{script_revision_id}/resource-slots/{slot_id}:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
      - {$ref: '#/components/parameters/SlotId'}
    put:
      tags: [Scripts]
      operationId: bindScriptResourceSlot
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResourceSlotBind'}
      responses:
        '200':
          description: Updated slot binding.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ResourceSlot'}
        '409': {$ref: '#/components/responses/Conflict'}
  /chapter-script-revisions/{script_revision_id}/vn-graph-compilations:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
    post:
      tags: [VNGraphs]
      operationId: compileVNGraph
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VNGraphCompileRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
  /chapter-script-revisions/{script_revision_id}/vn-graph-revisions:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
    get:
      tags: [VNGraphs]
      operationId: listVNGraphRevisions
      responses:
        '200':
          description: Graphs compiled from one exact Script IR.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/VNGraphRevision'}
  /chapter-script-revisions/{script_revision_id}/vn-graph-head:
    parameters:
      - {$ref: '#/components/parameters/ScriptRevisionId'}
    get:
      tags: [VNGraphs]
      operationId: getVNGraphHead
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [VNGraphs]
      operationId: updateVNGraphHead
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeadUpdate'}
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '409': {$ref: '#/components/responses/Conflict'}
  /vn-graph-revisions/{vn_graph_revision_id}:
    parameters:
      - {$ref: '#/components/parameters/VNGraphRevisionId'}
    get:
      tags: [VNGraphs]
      operationId: getVNGraphRevision
      responses:
        '200':
          description: Immutable compiled graph.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VNGraphRevision'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    ChapterRevisionId:
      name: chapter_revision_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    ScriptRevisionId:
      name: script_revision_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    VNGraphRevisionId:
      name: vn_graph_revision_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    SlotId:
      name: slot_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current integer lock version, optionally quoted.
      schema: {type: string, pattern: '^("[1-9][0-9]*"|[1-9][0-9]*)$'}
  responses:
...
    RevisionHead:
      description: Current reviewed revision pointer.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/RevisionHead'}
    TaskAccepted:
      description: Durable asynchronous task accepted.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TaskAccepted'}
...
    NotFound:
      description: Resource is absent or not visible to this caller.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Lock, idempotency, provenance, or source fingerprint conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    GenerationRequest:
      type: object
      properties:
        instructions: {type: [string, 'null'], maxLength: 12000}
        parameters: {type: object, additionalProperties: true}
...
    RevisionHead:
      type: object
      required: [revision_id, lock_version]
      properties:
        revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        lock_version: {type: integer, minimum: 1}
        updated_at: {type: string, format: date-time}
    HeadUpdate:
      type: object
      required: [revision_id]
      properties:
        revision_id: {$ref: '#/components/schemas/UUID'}
...
    ScriptRevision:
      type: object
      required: [id, chapter_revision_id, revision_no, source_hash, script_hash, script_json]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        chapter_revision_id: {$ref: '#/components/schemas/UUID'}
        revision_no: {type: integer, minimum: 1}
        source_hash: {type: string}
        script_hash: {type: string}
        script_json: {type: object, additionalProperties: true}
    ResourceSlot:
      type: object
      required: [id, script_revision_id, role, lock_version]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        script_revision_id: {$ref: '#/components/schemas/UUID'}
        role: {type: string}
        asset_version_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        lock_version: {type: integer, minimum: 1}
    ResourceRenderRequest:
      type: object
      required: [role]
      properties:
        role: {type: string}
    ResourceSlotBind:
      type: object
      required: [asset_version_id]
      properties:
        asset_version_id: {$ref: '#/components/schemas/UUID'}
    VNGraphCompileRequest:
      type: object
      properties:
        compiler_version: {type: [string, 'null']}
        schema_version: {type: [string, 'null']}
    VNGraphRevision:
      type: object
      required: [id, script_revision_id, binding_manifest, binding_manifest_hash, graph_hash, graph_json]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        script_revision_id: {$ref: '#/components/schemas/UUID'}
        binding_manifest: {type: object, additionalProperties: true}
        binding_manifest_hash: {type: string}
        graph_hash: {type: string}
        graph_json: {type: object, additionalProperties: true}
...
    TaskAccepted:
      type: object
      required: [task_id, status, events_url, created]
      properties:
        task_id: {$ref: '#/components/schemas/UUID'}
        status: {type: string}
        events_url: {type: string}
        created: {type: boolean}
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}